Move enum into types folder and update projects import
